requirements.txt
Flask==2.3.3
pandas==2.0.3
numpy==1.24.4
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
gunicorn